+++ apps/io-func-sign-issuer/src/infra/http/handlers/__test__/set-signature-request-status.spec.ts
@@ -124,6 +124,7 @@ describe("SetSignatureRequestHandler", () => {
       input: req,
       inputDecoder: H.HttpRequest,
       eventAnalyticsClient: {} as EventHubProducerClient,
+      legacyEventAnalyticsClient: {} as EventHubProducerClient,
       ready: {} as QueueClient,
       updated: {} as QueueClient,
     });
@@ -157,6 +158,7 @@ describe("SetSignatureRequestHandler", () => {
       input: req,
       inputDecoder: H.HttpRequest,
       eventAnalyticsClient: {} as EventHubProducerClient,
+      legacyEventAnalyticsClient: {} as EventHubProducerClient,
       ready: {} as QueueClient,
       updated: {} as QueueClient,
     });
@@ -190,6 +192,7 @@ describe("SetSignatureRequestHandler", () => {
       input: req,
       inputDecoder: H.HttpRequest,
       eventAnalyticsClient: {} as EventHubProducerClient,
+      legacyEventAnalyticsClient: {} as EventHubProducerClient,
       ready: {} as QueueClient,
       updated: {} as QueueClient,
     });
@@ -226,6 +229,7 @@ describe("SetSignatureRequestHandler", () => {
       input: req,
       inputDecoder: H.HttpRequest,
       eventAnalyticsClient: {} as EventHubProducerClient,
+      legacyEventAnalyticsClient: {} as EventHubProducerClient,
       ready: {
         sendMessage: (_: string) => Promise.resolve({}),
       } as QueueClient,
@@ -260,6 +264,7 @@ describe("SetSignatureRequestHandler", () => {
       input: req,
       inputDecoder: H.HttpRequest,
       eventAnalyticsClient: {} as EventHubProducerClient,
+      legacyEventAnalyticsClient: {} as EventHubProducerClient,
       ready: {} as QueueClient,
       updated: {
         sendMessage: (_: string) => Promise.resolve({}),