+++ apps/io-func-sign-issuer/src/upload.ts
@@ -5,7 +5,6 @@ import * as t from "io-ts";
 
 import * as RTE from "fp-ts/lib/ReaderTaskEither";
 import * as TE from "fp-ts/lib/TaskEither";
-import * as R from "fp-ts/lib/Reader";
 import * as E from "fp-ts/lib/Either";
 import * as O from "fp-ts/lib/Option";
 
@@ -123,7 +122,7 @@ export type FileStorage = {
     filename: string
   ) => TE.TaskEither<Error, string>;
   remove: (filename: string) => TE.TaskEither<Error, void>;
-  getUrl: (filename: string) => string;
+  getUrl: (filename: string) => TE.TaskEither<Error, string>;
 };
 
 export const getUploadedDocument =
@@ -172,6 +171,12 @@ export const createDocumentFromUrl =
     storage.createFromUrl(url, filename);
 
 export const getUploadedDocumentUrl =
-  (uploadId: string): R.Reader<{ uploadedFileStorage: FileStorage }, string> =>
+  (
+    uploadId: string
+  ): RTE.ReaderTaskEither<
+    { uploadedFileStorage: FileStorage },
+    Error,
+    string
+  > =>
   ({ uploadedFileStorage: storage }) =>
     storage.getUrl(uploadId);