+++ apps/io-func-sign-user/src/infra/http/handlers/get-third-party-message-attachments-content.ts
@@ -8,14 +8,14 @@ import { lookup } from "fp-ts/lib/Record";
 import { sequenceS } from "fp-ts/lib/Apply";
 
 import { Database } from "@azure/cosmos";
-import { ContainerClient } from "@azure/storage-blob";
+import { BaseContainerClientWithFallback } from "@pagopa/azure-storage-migration-kit";
 
 import { PdvTokenizerClientWithApiKey } from "@io-sign/io-sign/infra/pdv-tokenizer/client";
 import { makeGetSignerByFiscalCode } from "@io-sign/io-sign/infra/pdv-tokenizer/signer";
 import { EntityNotFoundError } from "@io-sign/io-sign/error";
 import { Document, DocumentId, DocumentReady } from "@io-sign/io-sign/document";
 import { GetDocumentContent } from "@io-sign/io-sign/document-content";
-import { getDocumentContent as getDocumentContentFromStorage } from "@io-sign/io-sign/infra/azure/storage/document-content";
+import { getDocumentContentWithFallback } from "@io-sign/io-sign/infra/azure/storage/blob-storage-with-fallback";
 import { bufferResponse } from "@io-sign/io-sign/infra/http/response";
 import { logErrorAndReturnResponse } from "@io-sign/io-sign/infra/http/utils";
 
@@ -39,7 +39,7 @@ const requireDocumentId = (
 export type GetThirdPartyMessageAttachmentContentDependencies = {
   pdvTokenizerClient: PdvTokenizerClientWithApiKey;
   db: Database;
-  signedContainerClient: ContainerClient;
+  signedContainerClient: BaseContainerClientWithFallback;
 };
 
 export const GetThirdPartyMessageAttachmentContentHandler = H.of(
@@ -63,10 +63,7 @@ export const GetThirdPartyMessageAttachmentContentHandler = H.of(
             const getDocumentContent: GetDocumentContent = (
               document: DocumentReady
             ) =>
-              pipe(
-                document,
-                getDocumentContentFromStorage
-              )(signedContainerClient);
+              getDocumentContentWithFallback(document)(signedContainerClient);
             const getSignedDocumentContent =
               makeGetSignedDocumentContent(getDocumentContent);