+++ infra/bootstrapper/prod/locals.tf
@@ -44,11 +44,14 @@ locals {
   }
 
   functions = {
-    itn_io_sign_rg_name           = "${local.itn_prefix}-sign-rg-01"
     io_p_sign_integration_rg_name = "${local.weu_prefix}-sign-integration-rg"
-    io_p_sign_data_rg_name          = "${local.weu_prefix}-sign-data-rg"
-    io_p_sign_sec_rg_name           = "${local.weu_prefix}-sign-sec-rg"
+    io_p_sign_data_rg_name        = "${local.weu_prefix}-sign-data-rg"
+    io_p_sign_sec_rg_name         = "${local.weu_prefix}-sign-sec-rg"
   }
+
+  keyvault_common_ids = [
+    data.azurerm_key_vault.common.id
+  ]
 
   tf_storage_account = {
     name                = "iopitntfst001"