+++ apps/io-func-sign-issuer/CHANGELOG.md
@@ -1,5 +1,61 @@
 # io-func-sign-issuer
 
+## 1.6.0
+
+### Minor Changes
+
+- bcbd02c: send event to Event Hub in ITN and WEU
+
+### Patch Changes
+
+- Updated dependencies [bcbd02c]
+  - @io-sign/io-sign@1.10.0
+
+## 1.5.0
+
+### Minor Changes
+
+- 7b64d31: switch traffic to itn signed documents container with weu fallback
+
+### Patch Changes
+
+- Updated dependencies [7b64d31]
+  - @io-sign/io-sign@1.9.0
+
+## 1.4.0
+
+### Minor Changes
+
+- fb57d04: add a method to truncate the message subject if it is longer than 121 characters
+
+### Patch Changes
+
+- Updated dependencies [fb57d04]
+  - @io-sign/io-sign@1.8.0
+
+## 1.3.0
+
+### Minor Changes
+
+- 2593e4c: Switch validated-documents container to ITN with migration-kit
+
+### Patch Changes
+
+- Updated dependencies [2593e4c]
+  - @io-sign/io-sign@1.7.0
+
+## 1.2.3
+
+### Patch Changes
+
+- 82c8a1b: Removing uploaded-document WEU trigger
+
+## 1.2.2
+
+### Patch Changes
+
+- c81b660: Init switch to ITN for uploaded-documents with double trigger (WEU - ITN)
+
 ## 1.2.1
 
 ### Patch Changes